Extract PrepareL2ToTunnel functions
- Moved PrepareL2ToTunnelV4 and PrepareL2ToTunnelV6 functions to
  a separate source file, to facilitate unit testing.
